fix 'make deploy'; enhance pipeline controller
This commit introduces kustomization resources in the `config/`
directory so that `make deploy` actually works. I forgot to add those
in the skeleton commit. Those config files are generated by
kubebuilder when initializing a project so we should just take them as
a start and optimize as we see fit going forward.

Also, the pipeline controller now actually tries to grab HelmReleases
from all referenced clusters in a given pipeline, logs it and sets the
Pipeline's status accordingly.

Also:

bump kustomize to 4.5.7

This should fix issues on M1 machines (darwin/arm64) as older version
of kustomize aren't available for those.

Co-authored-by: Yiannis <8741709+yiannistri@users.noreply.github.com>