Add environment promotion webhook server
A new web server is spun up as part of the manager binary that serves
as a webhook to initiate an environment promotion. It serves requests
at /promotion/{namespace}/{name}/{environment} and consumes
[events.Events](https://pkg.go.dev/github.com/fluxcd/pkg/runtime/events#Event).

It compares the event from the request payload with the Pipeline
resource identified by the request path and calls the injected
promotion strategy implementation. The only strategy implemented right
now is a noop.

Successful execution of the strategy leads to a 201 or 204 HTTP status
code.

Minor changes:

* align logging with Flux conventions

closes #30