use go-git-providers for creating/updating promotion PRs
This change removes our direct dependency on the
github.com/google/go-github library and exchanges it with a ggp-only
solution now that ggp has PR editing support.

This opens up the opportunity to support more than one git provider in
the future based on some input, e.g. the repo URL in the pipeline spec
or an explicit identifier in the CRD.

In order for us not having to maintain manual mocks of ggp interfaces
I introduced mockgen to our tooling.