only run update-chart workflow for a tagged image
The `registry_package` event is emitted every time an image is pushed
to the registry which led to the workflow being kicked off 3 times,
for the x86 image, for the ARM image and another time for the
multi-arch image.

By checking for an empty tag name we make sure the workflow only runs
once and only for the event that has the image tag.