feat: Add rate limiting to promotion webhook
* Rate limiter under pkg in case we want to reuse
* Log if a request hit rate limit
* Add `Retry-After` header on hitting rate limit

```json
{
  "level": "error",
  "ts": "2022-12-21T12:37:00.284+0100",
  "logger": "promotion",
  "msg": "rate limit hit",
  "kind": "promotion webhook rate limiter",
  "ip": "127.0.0.1",
  "error": "127.0.0.1 has reached max requests 5"
}
```

Considered approaches:

1. Use ahmedash95/ratelimit as an external library

Dropped that idea, because the repo is not maintained. The idea is good, with a
very simple rate limiter. I don't like the API to specify limits with a string.
I decided to use it as a base and add the logic under `pkg/` with a bit of
refactoring instead of using it.

2. Tried to find other rate limiting implementations

After a while I decided to now research more as it consumed already more time
than simply create a new one. All solutions I found fell under these three
categories:

a) too complex
b) not maintained
c) does not fit our needs (like heavily depends on Gin)

Closes #32

References:
* https://github.com/weaveworks/pipeline-controller/issues/32
* https://github.com/ahmedash95/ratelimit
* https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Retry-After