feat: add retries to the promotion webhook
The `retry` package has one algorithm now (exponential), but it can be
extended to have more options (like constant delay).

The the server has three new `retry` related CLI flags:
* `--promotion-retry-delay`: Delay between promotion retries in seconds. (default 3)
* `--promotion-retry-max-delay`: Maximum delay between promotion retries. (default 20)
* `--promotion-retry-threshold`: How many times a promotion should be retried. (default 3)

Closes #58

Why didn't I use an existing library for retries
------------------------------------------------

The main reason is, retry libraries are built around the concept of
retrying on client side for example http requests. Exponential retry
logic is good and I think that's the most reasonable one to use (for now
at least), but exponential can grow... well exponentially, but we may
want to limit that value, so the first `N` tries follow the exponential
delay, but we can be sure we don't have 2 minutes long delay between
retries, but we still want to retry `N+M` times, and we want to wait
maximum X seconds between retries after `N` tries
(for example: 2, 4, 8, 16, 20, 20, 20).

For example I tried avast/retry-go but it was a mess with DelayType to
implement the logic described above.

References:
* https://github.com/weaveworks/pipeline-controller/issues/58
* https://github.com/avast/retry-go