fix: flaky controller tests with shared memory
That was a fun ride. The error message said (`pipeline_controller_test.go:71`):

```
Expected
    <string>: Failed to set status for pipeline kube-test-vq4sh/pipeline-dhnc5: GitopsCluster.gitops.weave.works "wrong-cluster" not found; requeue
to contain substring
    <string>: Updated pipeline
```

Tried to figure out why, but then I saw the previous test case where the
expected error message was exactly this error (with different name). And
that showed me it has to be a collision between tests, something that's
used by tests and shared the same resources.

Yes that's the `testEventRecorder`, with a naive approach originally I
wrote that data structure as a simple slice and called reset at the
beginning of all test cases, but it is possible to get extra events from
previous objects as they are not destroyed after each test cases. Flow:

1. Test A starts.
2. Test A resets `testEventRecorder`.
3. Test A create `ObjectA`.
4. `envtest` emits an event for `ObjectA`.
5. Test A checks events and finds events for `ObjectA`, it's happy.
6. Test A is done.
7. Test B starts.
8. Test B resets `testEventRecorder`.
9. Test B create `ObjectB`.
10. `envtest` emits an event for `ObjectA` in the background.
10. `envtest` emits an event for `ObjectB` in the background.
11. Test A checks events and finds events for `ObjectA`, it's not happy.

It wound be logical to not check `events[0]`, and just say "there has to
be an event in that array with that message", but because it's
asynchronous, we use `gomega.Eventually` to wait until an event is
emitted, but that approach fails if the slice is not empty because of a
different object.

That leads to this patch. This new approach uses
`map[string][]testEvent` instead of a flat `[]testEvent` container,
where the string key is generated from the object metadata as
`{namespace}-{name}`. As we generate new pipeline objects in test case,
this will isolate the events, removes the need to `Reset()` the
container which could lead to a race condition without locks if tests
are executed parallel.

Closes #140

References:
* https://github.com/weaveworks/pipeline-controller/actions/runs/3837492080/jobs/6532856984
* https://github.com/weaveworks/pipeline-controller/issues/140