Fork controller code to controllers/leveltriggered
I want the new behaviour to be guarded by a feature flag, and rather
than sprinkling conditionals through the controller code, I'd prefer
to have a completely separate set of code and dispatch to it in
main.go.

A reasonable starting point is the pretty minimal existing controller
code. So: copy the controller code into a new package and adjust as
needed to get

    go test -v ./controllers/leveltriggered

to succeed.

This commit also makes the test main wait for the manager to shut
down. This is minorly superstitious on my part: my experience is that
failure to shut managers down cleanly can cause problems with
subsequent test runs.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>