Record target statuses in pipeline status
The level-triggered promotion algorithm will need to know the status
of each target in the pipeline, so it can make a decision on if and
where a promotion needs to be triggered. To be explicit about how it
makes that decision, that controller should record the status it's
seen for each target. This will help with testing, and makes the
Pipeline object self-contained -- you should be able to see from the
status where the pipeline is up to.

This commit introduces a couple more types into the API to represent
the status of a target. Each one gets a full name (the type,
namespace, name, and cluster if it's in another cluster).

When the controller goes through the environments, it fetches the
targets apps. For the sake of not making too many decisions at once,
I've implemented only the case of HelmRelease objects in the local
cluster. Once there are tests demonstrating it works, it can be
generalised to other app types (the API accepts _any_ API group,
version, and kind).

The test involving a GitOpsCluster is skipped, because we would now
expect to fetch the application object from the remote cluster it
represents, rather than just examine the cluster object itself.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>