Change the meaning of "Ready" to "all targets read"
To date, the controller didn't have much to do other than validate
that it could see the cluster objects mentioned in a pipeline object
(which wasn't itself necessary, since taking actions relied on the
information given in webhooks). Its idea of "Ready" was that it could
see all the clusters, and they were themselves ready. If any cluster
was not present, it could exit early as unready.

Now, however, we want to actually go and fetch the status of the
targets mentioned in the pipeline. "Ready" must have a different
meaning: it means the controller can reach all the targets, and
thereby determine the state of play for the pipeline. It could exit
early in principle, but I have preferred to assess all the targets
even if some are unavailable, and give a summary in the ready
condition at the end.

The additional test will fail for now, since it checks that adding a
previously missing app object will result in a ready pipeline, but
there's nothing to trigger the pipeline to be re-examined, yet.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>