Queue a pipeline when its target app is updated
This adds indexing to the level-triggered controller, and watches
`HelmRelease` objects. If a HelmRelease object changes, it's looked up
against the index, and any pipeline using it will be queued to be
reconciled.

This is enough to pass the unit test that checks if a Pipeline becomes
ready if it can see all its targets.

The indexing is general with respect to the application kind, because
it includes the Group and Kind in the index key. But:

 - the watch is for the HelmRelease type specifically; and,
 - it does not account for remote clusters.

The next aim will be to make scaffolding that watches any application,
when it's mentioned in a pipeline. Remote clusters will come later.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>