Make room for remote cluster clients
Eventually, this code will need to be able to query for targets in
remote clusters. To make room for that, I've done a small
rearrangement here such that the client used for a target is obtained
from a procedure, rather than assumed to be the client for the local
cluster. This change is more to check this structure makes sense, than
to alter the behaviour.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>