Verify that a target reflects its app's status
This commit expands the test a little to make sure that when an app
(HelmRelease, here) has its status updated, the pipeline's status is
updated in sympathy.

In making the test pass, this corrects a bug with the implementation
so far: when going through the environments in the controller, the
environment status was not entered into the map of environment
statuses.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>