Use `NewPipelineReconciler` in tests
The tests make their own `PipelineReconciler` struct, but there's a
constructor `NewPipelineReconciler` which could do some
initialisation, and that will missed. So: use the constructor in
suite-test.go.

However! This now fails to initialise the eventRecorder, leaving it to
default in `SetupWithManager`, and capturing events is part of the
tests.

I don't want to disturb existing behaviour*, so I've kept this change
to the level-triggered controller.

*How does it change? The event recorder is constructed using a
notification-controller URL and used _only_ for promotion
notifications. The event recorder for things that happen in the
reconciler is left to default, which means it won't send those to the
notification-controller. In practice it's unlikely to matter, since
any alerts set up should filter for the promotion events.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>