Update controller-runtime to v0.15.1
This is not quite the most recent version of controller-runtime, but
it's recent enough, and it's the version compatible with
sig.k8s.io/cluster-api, which I would like to introduce. In newer
releases, there are some problems with kubeyaml's openapi package that
I don't want to deal with right now.

The changes to patch up are:

 - the builder package uses `Watch(client.Object, ...)` rather than
   `Watch(source.Source, ...)`
 - `handler.MapFunc` now takes a `context.Context` as the first
   arguments, so you don't have to use `context.Background`

There's also some change to how Kubernetes events get stringified,
which broke a test (in notification_test.go).

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>