Garbage collect caches that aren't needed any more
It's possible that due to pipelines disappearing, or being updated,
some caches will no longer be needed. If these are not shut down, the
number of caches will only grow, which constitutes a leak of resources
(though not necessarily a serious one, since it will max out at
`clusters x types`).

To be able to shut down caches that are no longer needed, we need to
be able to do a few things:

 1. detect when they aren't needed
 2. stop them running when not needed
 3. stop them when the controller is shutting down

To do the first, I index the cache keys used by each pipeline. The
garbage collector regularly checks to see if each cache has entries in
the index; and if not, it's not used by any pipeline and can be shut
down.

To keep track of caches to consider for collection, the GC uses a
rate-limiting work queue. When the cache is created, it's put on the
queue; and each time it's considered and is still needed, it's
requeued with a longer retry, up to about eight minutes. This avoids
the question of finding an appropriate event to hook into, with the
downside of being a bit eventual.

The second and third things can be arranged by deriving contexts from
the manager's context. I have introduced `runner` (in runner.go) which
can be Start()ed by the manager and thus gain access to its context,
and which can then construct a context for each cache. Each cache gets
its own cancel func that can be used to shut it down, but will also be
shut down by the manager when it's shutting down itself.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>