Remove and ignore go workspace files
There's no watertight advice on whether to check go workspace files in
(go.work and go.work.sum), but generally they are regarded as
local. See https://github.com/golang/go/issues/53502 as the most
central discussion on this).

Here, they tend to get in the way, because `go.work.sum` is not
updated by `go mod tidy` and isn't subject to (the default) `-mod
readonly`. In the absence of a clear benefit to checking them in, I'm
removing them.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>