Avoid some duplication in CI run
The CI workflow runs both `make test` and `make docker-build`, and the
latter target already has `test` as a prerequisite. I'd rather be
explicit about what's run in workflows, so: remove `test` as a
prerequisite of `docker-build`. (It's not a prerequisite of `build`,
for the record.)

While I'm here, let's make the default target something useful: now it
runs `lint`, `test`, and `build`.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>