Fix flake in remote tests
The remote target tests check that a target app is marked as unready
with an error of "not found" if it doesn't exist. But, running from
scratch, first the target will be unready because the cache for that
cluster has not synced yet, _then_ it will be unready because it's not
found. When the tests are run locally, it goes through that first
state fast enough that you probably wouldn't notice it; but in hosted
test runs (i.e., in GitHub Actions), running a test env for the remote
server is taxing enough that the flake becomes observable.

Thus: use `Eventually` here to check that it reaches that state,
rather than asserting it straight away.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>