Check that a promotion isn't invoked repeatedly
If you've set up promotions via PR, or CI, or (hypothetical) chatbot,
there will be a human-scale lag between a promotion being triggered
and it taking effect in the target environment. If the controller
processes the pipeline during that time, it shouldn't trigger the
promotion again.

This commit expands the mock promotion machinery into a little state
machine, so the gap between promotion trigger and completion can be
explicit in test cases. It asserts that no promotion of {env, version}
will be attempted twice, which (_provided_ we don't roll any
environment back a version) is what we want to prove here.

There's a race between the cache getting an object with the promotion
status, and the reconciler processing the object again. It's possible
for the reconciler, having just triggered a promotion, to run again
and see a version of the pipeline object that doesn't have the
promotion recorded -- and that will mean attempting the promotion
again.

In testing, this happens almost always, because patches will cause the
object to be requeued before it's finished reconciling, and it'll run
again as soon as its exited. I would expect it to also be common in
normal operation; and the outcome is that a promotion is re-run, which
is not great!

To mitigate the race, I've added a predicate to the watch which will
drop updates that didn't change the spec. This means that the patches
earlier in Reconcile won't cause the object to be requeued, and it is
much, much less likely that it'll run again straight away. It's still
possible -- maybe the spec changed while it was running.

Other, more involved ways to avoid it:

 - use conflict detection on write to bail before running a promotion,
   if the observed pipeline has changed

 - create Promotion objects and guard promotions by trying to create
   (another form of detecting conflict on write)