Requeue when the pipeline isn't ready
If a cache has not yet synced, or a target is not to be found, and the
controller has incomplete information, then the safest thing to do is
not make any decisions on promotions, and instead requeue the
pipeline for another try later.

This fixes a problem that was exposed by adding a predicate to the
watch on Pipeline: if a cluster cache isn't ready when the controller
looks, and the target doesn't exist, there's nothing to trigger
another go-round. This presented itself as a problem with the remote
and Kustomization tests, which explicitly check for a "not found"
target status.

(I fixed another problem with those tests, in passing -- if they are
going to check for an updated target status, they had better fetch the
object again.)

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>