Shut down all test envs
As part of testing pipelines with targets in remote clusters, an
envtest Environment `leafEnv` is created to act as a remote
cluster. But it is left running after the tests, as one could see with
`ps` once `go test` has completed.

Since the controller is run in TestMain, the environment must be shut
down in TestMain -- i.e., not within an individual test. This commit
adds a simple mechanism for keeping track of envs to shut down after
the tests have run, and uses it both for the "main" cluster
environment and the leaf cluster.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>